requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
supabase
sentence-transformers
streamlit
pandas
facebook-business
